Aborts elegantly when the model is restarted for any reason. Issue #10
According to the current API you always have to start the model before
using it. If the service is restarted without calling /open, the calls
to the /detect endpoint respond with error 412 (Precondition Failed)

It should probably continue to work, but today it is our "protocol".
In the meantime, I must adapt Face Recognition for this latest change.