Change our protocol. Dont need call open to use the service. Issue #12
It is recommended, since the maximum size and the type of files are
reported there, but if not called (Because for example the service was
restarted, or Docker Swarm is being used), automatically initializes
models before using them.

The security model is the use of shared API keys. and in any case we
must reject the requests if they do not have the correct keys, or if
dont comply with the imposed restrictions.